dtypes for saving

+++ Simu_data/opengate/gaga_garf_optimize.py
@@ -181,24 +181,24 @@ def main():
 
                 print(f"[Epoch {epoch+1}/{n_epochs}] [Subset {subset+1}/{args.nsubsets}] Loss = {loss.item():8.4f}            ({time.time()-t0_epoch:.4f} s)")
 
-                proj_k = itk.image_from_array(output_projs.detach().cpu().numpy())
+                proj_k = itk.image_from_array(output_projs.float().detach().cpu().numpy())
                 proj_k.CopyInformation(measured_projections)
                 itk.imwrite(proj_k, os.path.join(args.output_folder, f"proj_{epoch + 1}_{subset+1}.mha"))
 
-                rec_k = itk.image_from_array(image_k_tensor.detach().cpu().numpy())
+                rec_k = itk.image_from_array(image_k_tensor.float().detach().cpu().numpy())
                 rec_k.CopyInformation(like_img)
                 itk.imwrite(rec_k, os.path.join(args.output_folder, f"rec_{epoch + 1}_{subset+1}.mha"))
 
-                sens_k = itk.image_from_array(simu.garf_detector.sensitivity_image.detach().cpu().numpy())
+                sens_k = itk.image_from_array(simu.garf_detector.sensitivity_image.float().detach().cpu().numpy())
                 sens_k.CopyInformation(like_img)
                 itk.imwrite(sens_k, os.path.join(args.output_folder, f"sens_{epoch + 1}_{subset+1}.mha"))
 
 
-            rec_k = itk.image_from_array(image_k_tensor.detach().cpu().numpy())
+            rec_k = itk.image_from_array(image_k_tensor.float().detach().cpu().numpy())
             rec_k.CopyInformation(like_img)
             itk.imwrite(rec_k, os.path.join(args.output_folder, f"rec_{epoch+1}.mha"))
 
-            output_projs_itk = itk.image_from_array(output_projs.detach().cpu().numpy())
+            output_projs_itk = itk.image_from_array(output_projs.float().detach().cpu().numpy())
             output_projs_itk.CopyInformation(measured_projections)
             itk.imwrite(output_projs_itk, os.path.join(args.output_folder, f"projs_{epoch+1}.mha"))
